# benchmark block cache size