--- a/enterprise/server/test/performance/cache/cache_test.go
+++ b/enterprise/server/test/performance/cache/cache_test.go
@@ -18,15 +18,17 @@ import (
 	"github.com/buildbuddy-io/buildbuddy/server/testutil/testenv"
 	"github.com/buildbuddy-io/buildbuddy/server/testutil/testfs"
 	"github.com/buildbuddy-io/buildbuddy/server/testutil/testport"
+	"github.com/buildbuddy-io/buildbuddy/server/util/compression"
 	"github.com/buildbuddy-io/buildbuddy/server/util/log"
 	"github.com/buildbuddy-io/buildbuddy/server/util/prefix"
 
+	repb "github.com/buildbuddy-io/buildbuddy/proto/remote_execution"
 	rspb "github.com/buildbuddy-io/buildbuddy/proto/resource"
 )
 
 const (
-	maxSizeBytes = int64(100000000) // 100MB
-	numDigests   = 100
+	maxSizeBytes = int64(1e9) // 100MB
+	numDigests   = 50
 )
 
 var (
@@ -56,6 +58,20 @@ func getAnonContext(t testing.TB, te *testenv.TestEnv) context.Context {
 type digestBuf struct {
 	d   *rspb.ResourceName
 	buf []byte
+}
+
+func makeCompressibleDigests(t testing.TB, numDigests int, digestSizeBytes int64) []*digestBuf {
+	digestBufs := make([]*digestBuf, 0, numDigests)
+	for i := 0; i < numDigests; i++ {
+		r, buf := testdigest.RandomCompressibleCASResourceBuf(t, digestSizeBytes, "")
+		r.Compressor = repb.Compressor_ZSTD
+		compressedBuf := compression.CompressZstd(nil, buf)
+		digestBufs = append(digestBufs, &digestBuf{
+			d:   r,
+			buf: compressedBuf,
+		})
+	}
+	return digestBufs
 }
 
 func makeDigests(t testing.TB, numDigests int, digestSizeBytes int64) []*digestBuf {
@@ -112,12 +128,22 @@ func getDistributedCache(t testing.TB, te *testenv.TestEnv, c interfaces.Cache) 
 	return dc
 }
 
-func getPebbleCache(t testing.TB, te *testenv.TestEnv) interfaces.Cache {
+func getPebbleCacheOptions(t testing.TB, blockSize int64) *pebble_cache.Options {
 	testRootDir := testfs.MakeTempDir(t)
-	pc, err := pebble_cache.NewPebbleCache(te, &pebble_cache.Options{
-		RootDirectory: testRootDir,
-		MaxSizeBytes:  maxSizeBytes,
-	})
+	activeKeyVersion := int64(5)
+	return &pebble_cache.Options{
+		RootDirectory:         testRootDir,
+		BlockCacheSizeBytes:   blockSize,
+		MaxSizeBytes:          maxSizeBytes,
+		ActiveKeyVersion:      &activeKeyVersion,
+		AverageChunkSizeBytes: 524288,
+		IncludeMetadataSize:   true,
+	}
+}
+
+func getPebbleCache(t testing.TB, te *testenv.TestEnv, blockSize int64) interfaces.Cache {
+	opts := getPebbleCacheOptions(t, blockSize)
+	pc, err := pebble_cache.NewPebbleCache(te, opts)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -128,9 +154,37 @@ func getPebbleCache(t testing.TB, te *testenv.TestEnv) interfaces.Cache {
 	return pc
 }
 
-func benchmarkSet(ctx context.Context, c interfaces.Cache, digestSizeBytes int64, b *testing.B) {
-	digestBufs := makeDigests(b, numDigests, digestSizeBytes)
+func getPebbleCacheWithDigests(ctx context.Context, t testing.TB, te *testenv.TestEnv, opts *pebble_cache.Options, dbufs []*digestBuf) interfaces.Cache {
+	pc, err := pebble_cache.NewPebbleCache(te, opts)
+	if err != nil {
+		t.Fatalf("failed to create the first pebble cache: %s", err)
+	}
+	err = pc.Start()
+	if err != nil {
+		t.Fatalf("failed to start the first pebble cache: %s", err)
+	}
+	setDigestsInCache(t, ctx, pc, dbufs)
+	err = pc.Stop()
+	if err != nil {
+		t.Fatalf("failed to stop the first pebble cache: %s", err)
+	}
 
+	// start a second cache with same options
+	pc2, err := pebble_cache.NewPebbleCache(te, opts)
+	if err != nil {
+		t.Fatalf("failed to create the second pebble cache: %s", err)
+	}
+	err = pc2.Start()
+	if err != nil {
+		t.Fatalf("failed to start the second pebble cache: %s", err)
+	}
+	t.Cleanup(func() {
+		pc2.Stop()
+	})
+	return pc2
+}
+
+func benchmarkSet(ctx context.Context, c interfaces.Cache, digestBufs []*digestBuf, b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
@@ -143,15 +197,13 @@ func benchmarkSet(ctx context.Context, c interfaces.Cache, digestSizeBytes int64
 	}
 }
 
-func benchmarkGet(ctx context.Context, c interfaces.Cache, digestSizeBytes int64, b *testing.B) {
-	digestBufs := makeDigests(b, numDigests, digestSizeBytes)
-	setDigestsInCache(b, ctx, c, digestBufs)
+func benchmarkGet(ctx context.Context, c interfaces.Cache, digestBufs []*digestBuf, b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
 		dbuf := digestBufs[rand.Intn(len(digestBufs))]
-		b.SetBytes(dbuf.d.GetDigest().GetSizeBytes())
+		b.SetBytes(int64(len(dbuf.buf)))
 		_, err := c.Get(ctx, dbuf.d)
 		if err != nil {
 			b.Fatal(err)
@@ -159,14 +211,12 @@ func benchmarkGet(ctx context.Context, c interfaces.Cache, digestSizeBytes int64
 	}
 }
 
-func benchmarkGetMulti(ctx context.Context, c interfaces.Cache, digestSizeBytes int64, b *testing.B) {
-	digestBufs := makeDigests(b, numDigests, digestSizeBytes)
-	setDigestsInCache(b, ctx, c, digestBufs)
+func benchmarkGetMulti(ctx context.Context, c interfaces.Cache, digestBufs []*digestBuf, b *testing.B) {
 	digests := make([]*rspb.ResourceName, 0, len(digestBufs))
 	var sumBytes int64
 	for _, dbuf := range digestBufs {
 		digests = append(digests, dbuf.d)
-		sumBytes += dbuf.d.GetDigest().GetSizeBytes()
+		sumBytes += int64(len(dbuf.buf))
 	}
 	b.ReportAllocs()
 	b.ResetTimer()
@@ -180,9 +230,7 @@ func benchmarkGetMulti(ctx context.Context, c interfaces.Cache, digestSizeBytes 
 	}
 }
 
-func benchmarkFindMissing(ctx context.Context, c interfaces.Cache, digestSizeBytes int64, b *testing.B) {
-	digestBufs := makeDigests(b, numDigests, digestSizeBytes)
-	setDigestsInCache(b, ctx, c, digestBufs)
+func benchmarkFindMissing(ctx context.Context, c interfaces.Cache, digestBufs []*digestBuf, b *testing.B) {
 	digests := make([]*rspb.ResourceName, 0, len(digestBufs))
 	for _, dbuf := range digestBufs {
 		digests = append(digests, dbuf.d)
@@ -205,77 +253,148 @@ type namedCache struct {
 }
 
 func getAllCaches(b *testing.B, te *testenv.TestEnv) []*namedCache {
-	dc := getDiskCache(b, te)
-	ddc := getDistributedCache(b, te, dc)
-	pc := getPebbleCache(b, te)
-	dpc := getDistributedCache(b, te, pc)
+	//dc := getDiskCache(b, te)
+	//ddc := getDistributedCache(b, te, dc)
+	//pc := getPebbleCache(b, te)
+	//dpc := getDistributedCache(b, te, pc)
 
 	time.Sleep(100 * time.Millisecond)
 	caches := []*namedCache{
-		{getMemoryCache(b), "Memory"},
-		{getDiskCache(b, te), "Disk"},
-		{ddc, "DDisk"},
-		{getPebbleCache(b, te), "Pebble"},
-		{dpc, "DPebble"},
+		//{getMemoryCache(b), "Memory"},
+		//{getDiskCache(b, te), "Disk"},
+		//{ddc, "DDisk"},
+		{getPebbleCache(b, te, 0), "Pebble"},
+		//{dpc, "DPebble"},
 	}
 	return caches
 }
 
+type namedPebbleOptions struct {
+	opts *pebble_cache.Options
+	Name string
+}
+
+func getAllPebbleCacheOptions(b *testing.B, te *testenv.TestEnv) []*namedPebbleOptions {
+	return []*namedPebbleOptions{
+		{getPebbleCacheOptions(b, 1e9), "blockSize=1e9"},
+		{getPebbleCacheOptions(b, 5e9), "blockSize=5e9"},
+		{getPebbleCacheOptions(b, 1e10), "blockSize=1e10"},
+		{getPebbleCacheOptions(b, 5e10), "blockSize=5e10"},
+	}
+}
+
 func BenchmarkSet(b *testing.B) {
-	sizes := []int64{10, 100, 1000, 10000}
+	sizes := []int64{10, 100, 1000, 10000, 1e5, 1e6, 1e7, 5e7, 1e8}
 	te := testenv.GetTestEnv(b)
 	ctx := getAnonContext(b, te)
 
 	for _, cache := range getAllCaches(b, te) {
 		for _, size := range sizes {
-			name := fmt.Sprintf("%s%d", cache.Name, size)
+			name := fmt.Sprintf("%ssize=%d", cache.Name, size)
 			b.Run(name, func(b *testing.B) {
-				benchmarkSet(ctx, cache, size, b)
+				digestBufs := makeCompressibleDigests(b, numDigests, size)
+				benchmarkSet(ctx, cache, digestBufs, b)
 			})
 		}
 	}
 }
 
 func BenchmarkGet(b *testing.B) {
-	sizes := []int64{10, 100, 1000, 10000}
+	sizes := []int64{10, 100, 1000, 10000, 1e5, 1e6, 1e7, 5e7, 1e8}
 	te := testenv.GetTestEnv(b)
 	ctx := getAnonContext(b, te)
 
 	for _, cache := range getAllCaches(b, te) {
 		for _, size := range sizes {
-			name := fmt.Sprintf("%s%d", cache.Name, size)
+			name := fmt.Sprintf("%s/size=%d", cache.Name, size)
 			b.Run(name, func(b *testing.B) {
-				benchmarkGet(ctx, cache, size, b)
+				digestBufs := makeCompressibleDigests(b, numDigests, size)
+				setDigestsInCache(b, ctx, cache, digestBufs)
+				benchmarkGet(ctx, cache, digestBufs, b)
+			})
+		}
+	}
+}
+
+func BenchmarkPebbleGet(b *testing.B) {
+	sizes := []int64{10, 100, 1000, 10000, 1e5, 1e6, 1e7, 5e7, 1e8}
+	te := testenv.GetTestEnv(b)
+	ctx := getAnonContext(b, te)
+	for _, namedOpts := range getAllPebbleCacheOptions(b, te) {
+		for _, size := range sizes {
+			name := fmt.Sprintf("%s/size=%d", namedOpts.Name, size)
+			b.Run(name, func(b *testing.B) {
+				digestBufs := makeCompressibleDigests(b, numDigests, size)
+				pc := getPebbleCacheWithDigests(ctx, b, te, namedOpts.opts, digestBufs)
+				benchmarkGet(ctx, pc, digestBufs, b)
 			})
 		}
 	}
 }
 
 func BenchmarkGetMulti(b *testing.B) {
-	sizes := []int64{10, 100, 1000, 10000}
+	sizes := []int64{10, 100, 1000, 10000, 1e5, 1e6, 1e7, 5e7, 1e8}
 	te := testenv.GetTestEnv(b)
 	ctx := getAnonContext(b, te)
 
 	for _, cache := range getAllCaches(b, te) {
 		for _, size := range sizes {
-			name := fmt.Sprintf("%s%d", cache.Name, size)
+			name := fmt.Sprintf("%s/size=%d", cache.Name, size)
 			b.Run(name, func(b *testing.B) {
-				benchmarkGetMulti(ctx, cache, size, b)
+				digestBufs := makeCompressibleDigests(b, numDigests, size)
+				setDigestsInCache(b, ctx, cache, digestBufs)
+				benchmarkGetMulti(ctx, cache, digestBufs, b)
+			})
+		}
+	}
+}
+
+func BenchmarkPebbleGetMulti(b *testing.B) {
+	sizes := []int64{10, 100, 1000, 10000, 1e5, 1e6, 1e7, 5e7, 1e8}
+	te := testenv.GetTestEnv(b)
+	ctx := getAnonContext(b, te)
+
+	for _, namedOpts := range getAllPebbleCacheOptions(b, te) {
+		for _, size := range sizes {
+			name := fmt.Sprintf("%s/size=%d", namedOpts.Name, size)
+			b.Run(name, func(b *testing.B) {
+				digestBufs := makeCompressibleDigests(b, numDigests, size)
+				pc := getPebbleCacheWithDigests(ctx, b, te, namedOpts.opts, digestBufs)
+				benchmarkGetMulti(ctx, pc, digestBufs, b)
 			})
 		}
 	}
 }
 
 func BenchmarkFindMissing(b *testing.B) {
-	sizes := []int64{10, 100, 1000, 10000}
+	sizes := []int64{10, 100, 1000, 10000, 1e5, 1e6, 1e7, 5e7, 1e8}
 	te := testenv.GetTestEnv(b)
 	ctx := getAnonContext(b, te)
 
 	for _, cache := range getAllCaches(b, te) {
 		for _, size := range sizes {
-			name := fmt.Sprintf("%s%d", cache.Name, size)
+			name := fmt.Sprintf("%s/size=%d", cache.Name, size)
 			b.Run(name, func(b *testing.B) {
-				benchmarkFindMissing(ctx, cache, size, b)
+				digestBufs := makeCompressibleDigests(b, numDigests, size)
+				setDigestsInCache(b, ctx, cache, digestBufs)
+				benchmarkFindMissing(ctx, cache, digestBufs, b)
+			})
+		}
+	}
+}
+
+func BenchmarkPebbleFindMissing(b *testing.B) {
+	sizes := []int64{10, 100, 1000, 10000, 1e5, 1e6, 1e7, 5e7, 1e8}
+	te := testenv.GetTestEnv(b)
+	ctx := getAnonContext(b, te)
+
+	for _, namedOpts := range getAllPebbleCacheOptions(b, te) {
+		for _, size := range sizes {
+			name := fmt.Sprintf("%s/size=%d", namedOpts.Name, size)
+			b.Run(name, func(b *testing.B) {
+				digestBufs := makeCompressibleDigests(b, numDigests, size)
+				pc := getPebbleCacheWithDigests(ctx, b, te, namedOpts.opts, digestBufs)
+				benchmarkFindMissing(ctx, pc, digestBufs, b)
 			})
 		}
 	}