<!-- Optional: Provide additional context (beyond the PR title). -->

<!-- Optional: link a GitHub issue.
     Example: "Fixes #123" will auto-close #123 when the PR is merged. -->

**Related issues**: N/A

benchmark on vm when pebble root dir is set on standard ssd:

```
goos: linux
goarch: amd64
cpu: Intel(R) Xeon(R) CPU @ 2.80GHz
                                    │     1e9      │                 5e9                  │                 1e10                 │                 5e10                 │
                                    │    sec/op    │    sec/op     vs base                │    sec/op     vs base                │    sec/op     vs base                │
PebbleGet/size=10-16                  14.56µ ± ∞ ¹   15.68µ ± ∞ ¹       ~ (p=1.000 n=1) ²   15.82µ ± ∞ ¹       ~ (p=1.000 n=1) ²   15.92µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100-16                 14.50µ ± ∞ ¹   15.79µ ± ∞ ¹       ~ (p=1.000 n=1) ²   15.88µ ± ∞ ¹       ~ (p=1.000 n=1) ²   16.14µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=1000-16                15.15µ ± ∞ ¹   16.24µ ± ∞ ¹       ~ (p=1.000 n=1) ²   16.08µ ± ∞ ¹       ~ (p=1.000 n=1) ²   16.31µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=10000-16               28.27µ ± ∞ ¹   28.44µ ± ∞ ¹       ~ (p=1.000 n=1) ²   29.04µ ± ∞ ¹       ~ (p=1.000 n=1) ²   28.28µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100000-16              31.59µ ± ∞ ¹   32.81µ ± ∞ ¹       ~ (p=1.000 n=1) ²   32.71µ ± ∞ ¹       ~ (p=1.000 n=1) ²   32.69µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=1000000-16             61.90µ ± ∞ ¹   61.27µ ± ∞ ¹       ~ (p=1.000 n=1) ²   60.45µ ± ∞ ¹       ~ (p=1.000 n=1) ²   60.66µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=10000000-16            768.6µ ± ∞ ¹   824.0µ ± ∞ ¹       ~ (p=1.000 n=1) ²   809.4µ ± ∞ ¹       ~ (p=1.000 n=1) ²   800.1µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=50000000-16            4.024m ± ∞ ¹   4.037m ± ∞ ¹       ~ (p=1.000 n=1) ²   4.348m ± ∞ ¹       ~ (p=1.000 n=1) ²   4.127m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100000000-16           8.247m ± ∞ ¹   8.258m ± ∞ ¹       ~ (p=1.000 n=1) ²   8.358m ± ∞ ¹       ~ (p=1.000 n=1) ²   8.217m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10-16             570.1µ ± ∞ ¹   565.3µ ± ∞ ¹       ~ (p=1.000 n=1) ²   561.3µ ± ∞ ¹       ~ (p=1.000 n=1) ²   575.5µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=100-16            571.7µ ± ∞ ¹   569.6µ ± ∞ ¹       ~ (p=1.000 n=1) ²   558.5µ ± ∞ ¹       ~ (p=1.000 n=1) ²   562.1µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=1000-16           576.2µ ± ∞ ¹   569.2µ ± ∞ ¹       ~ (p=1.000 n=1) ²   565.6µ ± ∞ ¹       ~ (p=1.000 n=1) ²   570.3µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10000-16          1.179m ± ∞ ¹   1.157m ± ∞ ¹       ~ (p=1.000 n=1) ²   1.176m ± ∞ ¹       ~ (p=1.000 n=1) ²   1.170m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=100000-16         1.221m ± ∞ ¹   1.211m ± ∞ ¹       ~ (p=1.000 n=1) ²   1.210m ± ∞ ¹       ~ (p=1.000 n=1) ²   1.214m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=1000000-16        1.495m ± ∞ ¹   1.506m ± ∞ ¹       ~ (p=1.000 n=1) ²   1.503m ± ∞ ¹       ~ (p=1.000 n=1) ²   1.508m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10000000-16       34.46m ± ∞ ¹   33.36m ± ∞ ¹       ~ (p=1.000 n=1) ²   33.64m ± ∞ ¹       ~ (p=1.000 n=1) ²   35.07m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=50000000-16       171.3m ± ∞ ¹   173.4m ± ∞ ¹       ~ (p=1.000 n=1) ²   170.3m ± ∞ ¹       ~ (p=1.000 n=1) ²   169.7m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=100000000-16      348.8m ± ∞ ¹   330.2m ± ∞ ¹       ~ (p=1.000 n=1) ²   344.7m ± ∞ ¹       ~ (p=1.000 n=1) ²   340.2m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=10-16          453.4µ ± ∞ ¹   446.9µ ± ∞ ¹       ~ (p=1.000 n=1) ²   448.6µ ± ∞ ¹       ~ (p=1.000 n=1) ²   442.9µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=100-16         453.0µ ± ∞ ¹   453.0µ ± ∞ ¹       ~ (p=1.000 n=1) ²   451.5µ ± ∞ ¹       ~ (p=1.000 n=1) ²   449.7µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=1000-16        463.6µ ± ∞ ¹   463.4µ ± ∞ ¹       ~ (p=1.000 n=1) ²   462.6µ ± ∞ ¹       ~ (p=1.000 n=1) ²   462.4µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=10000-16       462.0µ ± ∞ ¹   465.3µ ± ∞ ¹       ~ (p=1.000 n=1) ²   465.1µ ± ∞ ¹       ~ (p=1.000 n=1) ²   466.4µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=100000-16      465.5µ ± ∞ ¹   469.8µ ± ∞ ¹       ~ (p=1.000 n=1) ²   463.6µ ± ∞ ¹       ~ (p=1.000 n=1) ²   459.9µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=1000000-16     478.5µ ± ∞ ¹   470.1µ ± ∞ ¹       ~ (p=1.000 n=1) ²   467.9µ ± ∞ ¹       ~ (p=1.000 n=1) ²   464.0µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=10000000-16    3.066m ± ∞ ¹   3.055m ± ∞ ¹       ~ (p=1.000 n=1) ²   3.072m ± ∞ ¹       ~ (p=1.000 n=1) ²   3.006m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=50000000-16    14.09m ± ∞ ¹   14.02m ± ∞ ¹       ~ (p=1.000 n=1) ²   14.14m ± ∞ ¹       ~ (p=1.000 n=1) ²   13.91m ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=100000000-16   29.41m ± ∞ ¹   28.22m ± ∞ ¹       ~ (p=1.000 n=1) ²   30.05m ± ∞ ¹       ~ (p=1.000 n=1) ²   29.12m ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                               892.7µ         897.8µ        +0.56%                   902.7µ        +1.12%                   899.3µ        +0.74%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                 │      1e9      │                  5e9                  │                 1e10                  │                 5e10                  │
                                 │      B/s      │      B/s       vs base                │      B/s       vs base                │      B/s       vs base                │
PebbleGet/size=10-16               1.507Mi ± ∞ ¹   1.402Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.383Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.383Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100-16              1.383Mi ± ∞ ¹   1.268Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.259Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.240Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=1000-16             2.766Mi ± ∞ ¹   2.642Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   2.670Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   2.689Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=10000-16            6.380Mi ± ∞ ¹   6.409Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   5.751Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   6.104Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100000-16           31.59Mi ± ∞ ¹   29.91Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   30.12Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   30.39Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=1000000-16          256.5Mi ± ∞ ¹   186.5Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   230.7Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   263.3Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=10000000-16         240.5Mi ± ∞ ¹   215.8Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   217.6Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   206.2Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=50000000-16         231.1Mi ± ∞ ¹   239.8Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   221.6Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   232.1Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100000000-16        232.7Mi ± ∞ ¹   233.8Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   236.1Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   240.6Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10-16          1.926Mi ± ∞ ¹   1.936Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.955Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.907Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=100-16         1.755Mi ± ∞ ¹   1.755Mi ± ∞ ¹       ~ (p=1.000 n=1) ³   1.793Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.783Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=1000-16        3.452Mi ± ∞ ¹   3.586Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   3.576Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   3.490Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10000-16       7.391Mi ± ∞ ¹   7.515Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   7.391Mi ± ∞ ¹       ~ (p=1.000 n=1) ³   7.410Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=100000-16      40.65Mi ± ∞ ¹   40.88Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   41.01Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   40.84Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=1000000-16     515.9Mi ± ∞ ¹   506.2Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   499.5Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   500.3Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10000000-16    248.7Mi ± ∞ ¹   260.5Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   256.0Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   247.4Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=50000000-16    281.2Mi ± ∞ ¹   276.9Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   282.5Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   283.9Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=100000000-16   279.8Mi ± ∞ ¹   295.5Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   284.7Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   287.8Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                            28.31Mi         27.52Mi        -2.78%                   27.53Mi        -2.74%                   27.73Mi        -2.03%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal

                                    │      1e9      │                  5e9                  │                 1e10                  │                 5e10                  │
                                    │     B/op      │     B/op       vs base                │     B/op       vs base                │     B/op       vs base                │
PebbleGet/size=10-16                  4.229Ki ± ∞ ¹   4.220Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   4.222Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   4.221Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100-16                 4.229Ki ± ∞ ¹   4.220Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   4.221Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   4.223Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=1000-16                4.257Ki ± ∞ ¹   4.244Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   4.246Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   4.246Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=10000-16               5.355Ki ± ∞ ¹   5.338Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   5.339Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   5.338Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100000-16              8.011Ki ± ∞ ¹   7.999Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   7.997Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   8.001Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=1000000-16             78.93Ki ± ∞ ¹   76.54Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   75.22Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   75.81Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=10000000-16            1.402Mi ± ∞ ¹   1.440Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.427Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.431Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=50000000-16            7.695Mi ± ∞ ¹   7.705Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   7.698Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   7.683Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100000000-16           19.14Mi ± ∞ ¹   19.10Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   18.98Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   19.01Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10-16             154.1Ki ± ∞ ¹   154.1Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   154.1Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   154.2Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=100-16            154.1Ki ± ∞ ¹   154.1Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   154.1Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   154.2Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=1000-16           156.5Ki ± ∞ ¹   156.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   156.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   156.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10000-16          218.2Ki ± ∞ ¹   218.1Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   218.1Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   218.2Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=100000-16         267.5Ki ± ∞ ¹   267.3Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   267.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   267.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=1000000-16        1.083Mi ± ∞ ¹   1.070Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.057Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   1.065Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10000000-16       46.17Mi ± ∞ ¹   46.15Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   46.33Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   46.52Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=50000000-16       228.9Mi ± ∞ ¹   228.9Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   229.7Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   229.1Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=100000000-16      603.6Mi ± ∞ ¹   603.4Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   599.9Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   602.7Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=10-16          133.8Ki ± ∞ ¹   133.8Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   133.8Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   133.8Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=100-16         133.8Ki ± ∞ ¹   133.8Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   133.8Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   133.8Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=1000-16        135.0Ki ± ∞ ¹   134.9Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   135.0Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   135.0Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=10000-16       136.6Ki ± ∞ ¹   136.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   136.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   136.6Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=100000-16      136.5Ki ± ∞ ¹   136.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   136.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   136.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=1000000-16     136.6Ki ± ∞ ¹   136.6Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   136.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   136.6Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=10000000-16    869.5Ki ± ∞ ¹   869.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   870.7Ki ± ∞ ¹       ~ (p=1.000 n=1) ²   869.4Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=50000000-16    3.590Mi ± ∞ ¹   3.589Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   3.589Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   3.589Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=100000000-16   19.87Mi ± ∞ ¹   19.91Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   19.87Mi ± ∞ ¹       ~ (p=1.000 n=1) ²   20.11Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                               411.1Ki         410.6Ki        -0.11%                   410.0Ki        -0.26%                   410.5Ki        -0.13%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                    │     1e9      │                 5e9                  │                 1e10                 │                 5e10                 │
                                    │  allocs/op   │  allocs/op    vs base                │  allocs/op    vs base                │  allocs/op    vs base                │
PebbleGet/size=10-16                   85.00 ± ∞ ¹    85.00 ± ∞ ¹       ~ (p=1.000 n=1) ²    85.00 ± ∞ ¹       ~ (p=1.000 n=1) ²    85.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100-16                  85.00 ± ∞ ¹    85.00 ± ∞ ¹       ~ (p=1.000 n=1) ²    85.00 ± ∞ ¹       ~ (p=1.000 n=1) ²    85.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=1000-16                 85.00 ± ∞ ¹    85.00 ± ∞ ¹       ~ (p=1.000 n=1) ²    85.00 ± ∞ ¹       ~ (p=1.000 n=1) ²    85.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=10000-16                95.00 ± ∞ ¹    95.00 ± ∞ ¹       ~ (p=1.000 n=1) ²    95.00 ± ∞ ¹       ~ (p=1.000 n=1) ²    95.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=100000-16               103.0 ± ∞ ¹    103.0 ± ∞ ¹       ~ (p=1.000 n=1) ²    103.0 ± ∞ ¹       ~ (p=1.000 n=1) ²    103.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGet/size=1000000-16              137.0 ± ∞ ¹    136.0 ± ∞ ¹       ~ (p=1.000 n=1) ³    136.0 ± ∞ ¹       ~ (p=1.000 n=1) ³    136.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGet/size=10000000-16            1.179k ± ∞ ¹   1.180k ± ∞ ¹       ~ (p=1.000 n=1) ³   1.181k ± ∞ ¹       ~ (p=1.000 n=1) ³   1.181k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGet/size=50000000-16            5.255k ± ∞ ¹   5.259k ± ∞ ¹       ~ (p=1.000 n=1) ³   5.262k ± ∞ ¹       ~ (p=1.000 n=1) ³   5.262k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGet/size=100000000-16           68.98k ± ∞ ¹   68.65k ± ∞ ¹       ~ (p=1.000 n=1) ³   66.53k ± ∞ ¹       ~ (p=1.000 n=1) ³   67.16k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGetMulti/size=10-16             3.365k ± ∞ ¹   3.365k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.366k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.366k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGetMulti/size=100-16            3.365k ± ∞ ¹   3.366k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.366k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.366k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGetMulti/size=1000-16           3.366k ± ∞ ¹   3.366k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.366k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.366k ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleGetMulti/size=10000-16          3.868k ± ∞ ¹   3.868k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.869k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.870k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGetMulti/size=100000-16         3.875k ± ∞ ¹   3.875k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.876k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.877k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGetMulti/size=1000000-16        3.892k ± ∞ ¹   3.890k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.894k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.891k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGetMulti/size=10000000-16       57.27k ± ∞ ¹   57.28k ± ∞ ¹       ~ (p=1.000 n=1) ³   57.32k ± ∞ ¹       ~ (p=1.000 n=1) ³   57.35k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGetMulti/size=50000000-16       261.4k ± ∞ ¹   261.5k ± ∞ ¹       ~ (p=1.000 n=1) ³   261.5k ± ∞ ¹       ~ (p=1.000 n=1) ³   261.3k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleGetMulti/size=100000000-16      2.888M ± ∞ ¹   2.881M ± ∞ ¹       ~ (p=1.000 n=1) ³   2.814M ± ∞ ¹       ~ (p=1.000 n=1) ³   2.867M ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleFindMissing/size=10-16          3.012k ± ∞ ¹   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²
PebbleFindMissing/size=100-16         3.012k ± ∞ ¹   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.013k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleFindMissing/size=1000-16        3.012k ± ∞ ¹   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.013k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.013k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleFindMissing/size=10000-16       3.012k ± ∞ ¹   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.013k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleFindMissing/size=100000-16      3.012k ± ∞ ¹   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.013k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.013k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleFindMissing/size=1000000-16     3.012k ± ∞ ¹   3.012k ± ∞ ¹       ~ (p=1.000 n=1) ²   3.013k ± ∞ ¹       ~ (p=1.000 n=1) ³   3.013k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleFindMissing/size=10000000-16    18.94k ± ∞ ¹   18.94k ± ∞ ¹       ~ (p=1.000 n=1) ³   18.94k ± ∞ ¹       ~ (p=1.000 n=1) ³   18.94k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleFindMissing/size=50000000-16    79.14k ± ∞ ¹   79.14k ± ∞ ¹       ~ (p=1.000 n=1) ³   79.14k ± ∞ ¹       ~ (p=1.000 n=1) ³   79.15k ± ∞ ¹       ~ (p=1.000 n=1) ³
PebbleFindMissing/size=100000000-16   363.4k ± ∞ ¹   364.3k ± ∞ ¹       ~ (p=1.000 n=1) ³   363.4k ± ∞ ¹       ~ (p=1.000 n=1) ³   367.5k ± ∞ ¹       ~ (p=1.000 n=1) ³
geomean                               3.960k         3.959k        -0.04%                   3.951k        -0.23%                   3.957k        -0.08%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ need >= 4 samples to detect a difference at alpha level 0.05
```
The following results are from my own machine:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
…